+++ app/src/main/java/com/jin/jjinweather/MyApp.kt
@@ -1,7 +1,7 @@
 package com.jin.jjinweather
 
 import android.app.Application
-import com.jin.jjinweather.layer.data.database.DatabaseProvider
+import com.jin.jjinweather.feature.database.data.DatabaseProvider
 
 class MyApp : Application() {
     override fun onCreate() {
