--- app/src/main/java/com/jin/jjinweather/layer/domain/model/PermissionState.kt
@@ -1,5 +1,0 @@
-package com.jin.jjinweather.layer.domain.model
-
-enum class PermissionState {
-    GRANTED, DENIED, SHOW_RATIONALE
-}
